requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
streamlit
python-dotenv
google-generativeai
xgboost        
